// src/Interfaces/lid.ts
export interface ILid {
  lid: string,
  adDomain: string | null
  adPath: string | null
  domain: string | null
  campaignId: number | null
  subCampaign?: string | null
  cid?: string | null
  payin: number | null
  payout: number | null
  advertiserId: number | null
  advertiserManagerId: number | null
  advertiserName: string | null
  verticalId: number | null
  verticalName: string | null
  offerId: number | null
  offerName: string | null
  offerType: string | null
  offerDescription: string | null
  region: string | null
  isp: string | null
  IP: string | null
  sflServer: string | null
  userAgent: string | null
  browser: string | null
  browserEngine: string | null
  browserVersion: string | null
  affiliateId: number | null
  affiliateManagerId: number | null
  payoutPercent: number | null
  deviceType: string | null
  isCpmOptionEnabled: boolean | number
  landingPageUrl: string | null
  landingPageId: number | null
  conversionType: string | null
  os: string | null
  referer: string | null
  originPayIn: number | null
  originPayOut: number | null
  originAdvertiserId: number | null
  originAdvertiserName: string | null
  originConversionType: string | null
  originIsCpmOptionEnabled: boolean | number
  originOfferId: number | null
  originVerticalId: number | null
  originVerticalName: string | null
  landingPageIdOrigin: number | null
  landingPageUrlOrigin: string | null
  capOverrideOfferId: number | null
  offerIdRedirectExitTraffic: number | null
  redirectType: string | null
  redirectReason: string | null
  capsType: string | null
  country: string | null
  platform: string | null
  isUseDefaultOfferUrl: boolean | null
  nestedExitOfferSteps: string | null
  fingerPrintInfo: string | null
  fingerPrintKey: string | null
  isUniqueVisit: boolean | null
  eventType: string | null
  ttl?: number | null
  _messageType: string | null
}